moved get_det3 function and fixed calls
